make errorprone happy

+++ lucene/test-framework/src/java/org/apache/lucene/tests/index/BaseKnnVectorsFormatTestCase.java
@@ -1496,7 +1496,8 @@ public abstract class BaseKnnVectorsFormatTestCase extends BaseIndexFileFormatTe
           int size = Lucene99HnswVectorsReader.EXHAUSTIVE_BULK_SCORE_ORDS;
           assertTrue(
               visitedLimit == results.totalHits.value()
-                  || ((visitedLimit + size - 1) / size) * size == results.totalHits.value());
+                  || ((visitedLimit + size - 1) / size) * ((long) size)
+                      == results.totalHits.value());
 
           // check the limit is not hit when it clearly exceeds the number of vectors
           k = vectorValues.size();
